LayoutTests/js/dom/global-constructors-attributes-idb.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Test to ensure that IndexedDB global constructors have the right attributes");
</script>
<script src="script-tests/global-constructors-attributes-idb.js"></script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
