LayoutTests/js/dom/global-constructors-attributes-idb-expected.txt

Test to ensure that IndexedDB global constructors have the right attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Object.getOwnPropertyDescriptor(global, 'IDBCursor').value is IDBCursor
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursor').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursor').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursor').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursor').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').value is IDBCursorWithValue
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBCursorWithValue').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBDatabase').value is IDBDatabase
PASS Object.getOwnPropertyDescriptor(global, 'IDBDatabase').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBDatabase').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBDatabase').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBDatabase').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBFactory').value is IDBFactory
PASS Object.getOwnPropertyDescriptor(global, 'IDBFactory').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBFactory').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBFactory').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBFactory').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBIndex').value is IDBIndex
PASS Object.getOwnPropertyDescriptor(global, 'IDBIndex').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBIndex').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBIndex').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBIndex').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').value is IDBKeyRange
PASS Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBKeyRange').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').value is IDBObjectStore
PASS Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBObjectStore').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').value is IDBOpenDBRequest
PASS Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBOpenDBRequest').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBRequest').value is IDBRequest
PASS Object.getOwnPropertyDescriptor(global, 'IDBRequest').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBRequest').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBRequest').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBRequest').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBTransaction').value is IDBTransaction
PASS Object.getOwnPropertyDescriptor(global, 'IDBTransaction').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBTransaction').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBTransaction').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBTransaction').configurable is true
PASS Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').value is IDBVersionChangeEvent
PASS Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').hasOwnProperty('get') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').hasOwnProperty('set') is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').enumerable is false
PASS Object.getOwnPropertyDescriptor(global, 'IDBVersionChangeEvent').configurable is true
PASS successfullyParsed is true

TEST COMPLETE
